// contracts/generated/typescript/models/EvidenceSourceDetail.ts
export interface EvidenceSourceDetail {
    /**
     *
     * @type {string}
     * @memberof EvidenceSourceDetail
     */
    id: string;
    /**
     *
     * @type {EvidenceSourceDetailSourceTypeEnum}
     * @memberof EvidenceSourceDetail
     */
    sourceType: EvidenceSourceDetailSourceTypeEnum;
    /**
     *
     * @type {string}
     * @memberof EvidenceSourceDetail
     */
    title: string;
    /**
     *
     * @type {string}
     * @memberof EvidenceSourceDetail
     */
    author?: string;
    /**
     *
     * @type {string}
     * @memberof EvidenceSourceDetail
     */
    publisher?: string;
    /**
     *
     * @type {Date}
     * @memberof EvidenceSourceDetail
     */
    publishedAt?: Date;
    /**
     *
     * @type {{ [key: string]: any; }}
     * @memberof EvidenceSourceDetail
     */
    metadata: { [key: string]: any; };
    /**
     *
     * @type {ExternalResourceDetail}
     * @memberof EvidenceSourceDetail
     */
    externalResource?: ExternalResourceDetail;
}


/**
 * @export
 */
export const EvidenceSourceDetailSourceTypeEnum = {
    Webpage: 'webpage',
    Pdf: 'pdf',
    Book: 'book',
    Image: 'image',
    Video: 'video',
    Api: 'api',
    Database: 'database'
} as const;
export type EvidenceSourceDetailSourceTypeEnum = typeof EvidenceSourceDetailSourceTypeEnum[keyof typeof EvidenceSourceDetailSourceTypeEnum];