// contracts/generated/typescript/models/EvidenceSourceVersion.ts
export interface EvidenceSourceVersion {
    /**
     *
     * @type {string}
     * @memberof EvidenceSourceVersion
     */
    id: string;
    /**
     *
     * @type {string}
     * @memberof EvidenceSourceVersion
     */
    sourceId: string;
    /**
     *
     * @type {string}
     * @memberof EvidenceSourceVersion
     */
    versionHash: string;
    /**
     *
     * @type {string}
     * @memberof EvidenceSourceVersion
     */
    rawAssetId?: string;
    /**
     *
     * @type {string}
     * @memberof EvidenceSourceVersion
     */
    extractedAssetId?: string;
    /**
     *
     * @type {Date}
     * @memberof EvidenceSourceVersion
     */
    fetchedAt: Date;
    /**
     *
     * @type {Date}
     * @memberof EvidenceSourceVersion
     */
    createdAt: Date;
}